% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countOccurrences.R
\name{countOccurrences}
\alias{countOccurrences}
\title{This function counts the occurrences of specified concept ids within specified tables in an OMOP database,
including direct occurrences and occurrences through descendants, based on a provided cdm_schema and links.
It returns a tibble summarizing the counts across persons and across records.}
\usage{
countOccurrences(
  v,
  tables,
  links,
  db_connection,
  cdm_schema,
  vocab_schema,
  save_path = NULL
)
}
\arguments{
\item{v}{A vector of concept_ids to count occurrences for.}

\item{tables}{A character vector of CDM table names to search within.}

\item{links}{A list linking each table to its respective concept_id field.}

\item{db_connection}{A database connection object through which queries will be executed.}

\item{cdm_schema}{The database cdm_schema in which the tables are located.}

\item{vocab_schema}{The database vocab_schema in which the concept tables are located.}

\item{save_path}{The path to save the results to as a CSV file. If NULL, the results are not saved.}
}
\value{
A tibble with columns for the number of times any concept from 'v' occurs: direct count of persons,
        direct count of records, descendant count of persons, and descendant count of records.
        The tibble also includes the concept names and domain ids derived from `v` and is arranged by the total
        record count (direct + descendant).
}
\description{
This function counts the occurrences of specified concept ids within specified tables in an OMOP database,
including direct occurrences and occurrences through descendants, based on a provided cdm_schema and links.
It returns a tibble summarizing the counts across persons and across records.
}
