% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStandardDB.R
\name{isStandardDB}
\alias{isStandardDB}
\title{Identify Non-Standard Concepts in Database}
\usage{
isStandardDB(db_connection, cdm_schema, vocab_schema, links, save_path = NULL)
}
\arguments{
\item{db_connection}{A DBI database connection object to the SQL database containing the standard concepts table.}

\item{cdm_schema}{The schema name of the Common Data Model (CDM) database containing the source tables.}

\item{vocab_schema}{The schema name of the vocabulary database containing the standard concepts table.}

\item{links}{A named list where each name is a table name and each value is a vector of two strings: the column names for `concept_id` and `source_code` in that table.}

\item{save_path}{(Optional) The file path where the joined tables containing non-standard concepts should be saved. If not provided, the tables will not be saved but will still be checked for non-standard concepts.}
}
\value{
A tibble containing the columns `concept_id`, `concept_name`, `source_code`, and `source_table`, representing the concept ID, concept name, source code, and source table name for each non-standard concept found.
}
\description{
This function checks a set of tables for non-standard concepts by comparing them against the concept table in a SQL database. It identifies non-standard concepts and optionally saves the joined tables containing these non-standard concepts.
}
\details{
The function first queries the `concept` table from the provided SQL database to retrieve the standard concepts. It then reads each specified table, preparing and joining it with the standard concepts table based on the concept ID. Non-standard concepts are identified based on the `standard_concept` column being `NA`. Information about these non-standard concepts is collected and returned as a tibble. The function also adds a `source_table` column to indicate the source table name.

If a `save_path` is provided, each joined table (including both standard and non-standard concepts) is saved to the specified directory with the same name as the original table file.

non_standard_concepts <- isStandardDB(db_conn, "cdm_schema", "vocab_schema", links, "path/to/save_non_standard/")
}
